セキュリティ強化のため、レスポンスヘッダー追加、リクエストヘッダーチェックを追加 #278

+++ app/src/Web/Controller/Admin/FilesController.php
@@ -19,6 +19,10 @@ class FilesController extends AdminController
    */
   public function ajax_index(Request $request): string
   {
+    if ($this->isInvalidAjaxRequest($request)) {
+      return $this->error403();
+    }
+
     $files_model = new FilesModel();
 
     $blog_id = $this->getBlogId($request);
@@ -280,13 +284,17 @@ class FilesController extends AdminController
    */
   public function ajax_delete(Request $request): string
   {
+    if ($this->isInvalidAjaxRequest($request)) {
+      return $this->error403();
+    }
+
     // 削除処理
     $json = array('status' => 0);
     if (!Model::load('Files')->deleteByIdsAndBlogId($request->get('id'), $this->getBlogId($request))) {
       $json = array('status' => 1);
     }
 
-    $this->set('http_content_type', "application/json; charset=utf-8");
+    $this->setContentType("application/json; charset=utf-8");
     $this->set('json', $json);
     return "admin/common/json.twig";
   }
